Added purging database capability for wintersoil

+++ app/controllers/purge_controller.rb
@@ -1,8 +1,11 @@
 class PurgeController < ApplicationController
 
   def deleting
-    Message.delete_all
-    redirect_to root_path
+    if is_owner?
+      Message.delete_all
+      redirect_to root_path
+    else
+    end
   end
 
 end
